22/7

+++ server/SignIn/Tasks.cs
@@ -31,7 +31,7 @@ namespace SignIn
         public Nullable<System.DateTime> Deadline { get; set; }
         public bool isDone { get; set; }
         public bool isDeleted { get; set; }
-        public Nullable<int> PriceQuoteTime { get; set; }
+        public Nullable<double> PriceQuoteTime { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Activity> Activity { get; set; }
